更新publit和userinf,修复bug

+++ db/follow.go
@@ -9,25 +9,25 @@ import (
 )
 
 //FollowCountByID:根据用户id获取关注的数量
-func FollowCountByID(uid int) (int, error) {
+func FollowCountByID(uid int) (int64, error) {
 	user := &model.User{}
 	common.GetDB().Where("id = ?", uid).First(user)
 	if user.ID == 0 {
 		return -1, errors.New("no such user")
 	}
 	cnt := common.GetDB().Model(user).Association("Follows").Count()
-	return cnt, nil
+	return int64(cnt), nil
 }
 
 //FollowerCountByID:根据用户id获取关注者的数量
-func FollowerCountByID(uid int) (int, error) {
+func FollowerCountByID(uid int) (int64, error) {
 	user := &model.User{}
 	common.GetDB().Where("id = ?", uid).First(user)
 	if user.ID == 0 {
 		return -1, errors.New("no such user")
 	}
 	cnt := common.GetDB().Model(user).Association("Followers").Count()
-	return cnt, nil
+	return int64(cnt), nil
 }
 
 //判断是否已关注作者
